draft GAN trainer
