+++ use-cases/mnist/torch/model.py
@@ -7,11 +7,13 @@
 # - Matteo Bunino <matteo.bunino@cern.ch> - CERN
 # --------------------------------------------------------------------------------------
 
+import torch.nn as nn
 import torch.nn.functional as F
-from torch import nn
 
 
 class Net(nn.Module):
+    """Simple NN classifier for MNIST."""
+
     def __init__(self):
         super(Net, self).__init__()
         self.conv1 = nn.Conv2d(1, 10, kernel_size=5)
@@ -28,3 +30,64 @@ class Net(nn.Module):
         x = F.dropout(x, training=self.training)
         x = self.fc2(x)
         return F.log_softmax(x, dim=0)
+
+
+class Generator(nn.Module):
+    """Simple GAN generator for MNIST."""
+
+    def __init__(self, z_dim, g_hidden, image_channel=1):
+        super(Generator, self).__init__()
+        self.main = nn.Sequential(
+            # input layer
+            nn.ConvTranspose2d(z_dim, g_hidden * 8, 4, 1, 0, bias=False),
+            nn.BatchNorm2d(g_hidden * 8),
+            nn.ReLU(True),
+            # 1st hidden layer
+            nn.ConvTranspose2d(g_hidden * 8, g_hidden * 4, 4, 2, 1, bias=False),
+            nn.BatchNorm2d(g_hidden * 4),
+            nn.ReLU(True),
+            # 2nd hidden layer
+            nn.ConvTranspose2d(g_hidden * 4, g_hidden * 2, 4, 2, 1, bias=False),
+            nn.BatchNorm2d(g_hidden * 2),
+            nn.ReLU(True),
+            # 3rd hidden layer
+            nn.ConvTranspose2d(g_hidden * 2, g_hidden, 4, 2, 1, bias=False),
+            nn.BatchNorm2d(g_hidden),
+            nn.ReLU(True),
+            # output layer
+            nn.ConvTranspose2d(g_hidden, image_channel, 4, 2, 1, bias=False),
+            nn.Tanh(),
+        )
+
+    def forward(self, input):
+        return self.main(input)
+
+
+class Discriminator(nn.Module):
+    """Simple GAN discriminator for MNIST."""
+
+    def __init__(self, d_hidden, image_channel=1):
+        super(Discriminator, self).__init__()
+        self.main = nn.Sequential(
+            # 1st layer
+            nn.Conv2d(image_channel, d_hidden, 4, 2, 1, bias=False),
+            nn.LeakyReLU(0.2, inplace=True),
+            # 2nd layer
+            nn.Conv2d(d_hidden, d_hidden * 2, 4, 2, 1, bias=False),
+            nn.BatchNorm2d(d_hidden * 2),
+            nn.LeakyReLU(0.2, inplace=True),
+            # 3rd layer
+            nn.Conv2d(d_hidden * 2, d_hidden * 4, 4, 2, 1, bias=False),
+            nn.BatchNorm2d(d_hidden * 4),
+            nn.LeakyReLU(0.2, inplace=True),
+            # 4th layer
+            nn.Conv2d(d_hidden * 4, d_hidden * 8, 4, 2, 1, bias=False),
+            nn.BatchNorm2d(d_hidden * 8),
+            nn.LeakyReLU(0.2, inplace=True),
+            # output layer
+            nn.Conv2d(d_hidden * 8, 1, 4, 1, 0, bias=False),
+            nn.Sigmoid(),
+        )
+
+    def forward(self, input):
+        return self.main(input).view(-1, 1).squeeze(1)
